refactor : view, controller 분리
